forensics-toolkit/requirements.txt
streamlit==1.35.0
pandas==2.2.2
scikit-learn==1.4.2
Pillow==10.3.0
folium==0.16.0
streamlit-folium==0.16.0
requests==2.31.0
python-magic-bin==0.4.14
pymediainfo==7.0.1
